add rules tests
